Atualizando BD

+++ avaliacao_final/create_DataBase.sql
@@ -4,7 +4,7 @@ USE loja_virtual;
 
 -- Empresa
 CREATE TABLE company (
-  id BIGINT AUTO_INCREMENT PRIMARY KEY,
+  id INT AUTO_INCREMENT PRIMARY KEY,
   name VARCHAR(255) NOT NULL,
   cnpj VARCHAR(32) NOT NULL,
   email VARCHAR(255),
@@ -19,7 +19,7 @@ CREATE TABLE company (
 
 -- Cliente
 CREATE TABLE customer (
-  id BIGINT AUTO_INCREMENT PRIMARY KEY,
+  id INT AUTO_INCREMENT PRIMARY KEY,
   full_name VARCHAR(255) NOT NULL,
   cpf VARCHAR(16) NOT NULL,
   email VARCHAR(255) UNIQUE NOT NULL,
@@ -35,7 +35,7 @@ CREATE TABLE customer (
 
 -- Cartão de fidelidade (1:1 com customer)
 CREATE TABLE loyalty_card (
-  id BIGINT AUTO_INCREMENT PRIMARY KEY,
+  id INT AUTO_INCREMENT PRIMARY KEY,
   customer_id BIGINT UNIQUE NOT NULL,
   code VARCHAR(64) UNIQUE NOT NULL,
   points INT DEFAULT 0,
@@ -46,7 +46,7 @@ CREATE TABLE loyalty_card (
 
 -- Fornecedor
 CREATE TABLE supplier (
-  id BIGINT AUTO_INCREMENT PRIMARY KEY,
+  id INT AUTO_INCREMENT PRIMARY KEY,
   name VARCHAR(255) NOT NULL,
   cnpj VARCHAR(32) NOT NULL,
   email VARCHAR(255),
@@ -61,7 +61,7 @@ CREATE TABLE supplier (
 
 -- Categoria (com hierarquia)
 CREATE TABLE category (
-  id BIGINT AUTO_INCREMENT PRIMARY KEY,
+  id INT AUTO_INCREMENT PRIMARY KEY,
   name VARCHAR(255) NOT NULL,
   description TEXT,
   parent_category_id BIGINT,
@@ -70,7 +70,7 @@ CREATE TABLE category (
 
 -- Depósito
 CREATE TABLE warehouse (
-  id BIGINT AUTO_INCREMENT PRIMARY KEY,
+  id INT AUTO_INCREMENT PRIMARY KEY,
   name VARCHAR(255) NOT NULL,
   address_street VARCHAR(255),
   address_number VARCHAR(32),
@@ -82,7 +82,7 @@ CREATE TABLE warehouse (
 
 -- Produto
 CREATE TABLE product (
-  id BIGINT AUTO_INCREMENT PRIMARY KEY,
+  id INT AUTO_INCREMENT PRIMARY KEY,
   company_id BIGINT NOT NULL,
   sku VARCHAR(64) UNIQUE NOT NULL,
   name VARCHAR(255) NOT NULL,
@@ -99,7 +99,7 @@ CREATE TABLE product (
 
 -- Imagem de produto (atributo multivalorado)
 CREATE TABLE product_image (
-  id BIGINT AUTO_INCREMENT PRIMARY KEY,
+  id INT AUTO_INCREMENT PRIMARY KEY,
   product_id BIGINT NOT NULL,
   url TEXT NOT NULL,
   alt_text VARCHAR(255),
@@ -108,8 +108,8 @@ CREATE TABLE product_image (
 
 -- Relação N:N produto-categoria
 CREATE TABLE product_category (
-  product_id BIGINT NOT NULL,
-  category_id BIGINT NOT NULL,
+  product_id INT NOT NULL,
+  category_id INT NOT NULL,
   PRIMARY KEY (product_id, category_id),
   FOREIGN KEY (product_id) REFERENCES product(id) ON DELETE CASCADE,
   FOREIGN KEY (category_id) REFERENCES category(id) ON DELETE CASCADE
@@ -117,8 +117,8 @@ CREATE TABLE product_category (
 
 -- Relação N:N produto-fornecedor
 CREATE TABLE product_supplier (
-  product_id BIGINT NOT NULL,
-  supplier_id BIGINT NOT NULL,
+  product_id INT NOT NULL,
+  supplier_id INT NOT NULL,
   supplier_sku VARCHAR(64),
   cost_price DECIMAL(12,2),
   PRIMARY KEY (product_id, supplier_id),
@@ -129,8 +129,8 @@ CREATE TABLE product_supplier (
 -- Estoque (Inventory)
 CREATE TABLE inventory (
   id BIGINT AUTO_INCREMENT PRIMARY KEY,
-  product_id BIGINT NOT NULL,
-  warehouse_id BIGINT NOT NULL,
+  product_id INT NOT NULL,
+  warehouse_id INT NOT NULL,
   quantity_on_hand INT NOT NULL DEFAULT 0,
   quantity_reserved INT NOT NULL DEFAULT 0,
   UNIQUE (product_id, warehouse_id),
@@ -140,7 +140,7 @@ CREATE TABLE inventory (
 
 -- Pedido
 CREATE TABLE `order` (
-  id BIGINT AUTO_INCREMENT PRIMARY KEY,
+  id INT AUTO_INCREMENT PRIMARY KEY,
   customer_id BIGINT NOT NULL,
   order_number VARCHAR(64) UNIQUE NOT NULL,
   order_date DATE NOT NULL,
@@ -150,7 +150,7 @@ CREATE TABLE `order` (
 
 -- Item do pedido
 CREATE TABLE order_item (
-  id BIGINT AUTO_INCREMENT PRIMARY KEY,
+  id INT AUTO_INCREMENT PRIMARY KEY,
   order_id BIGINT NOT NULL,
   product_id BIGINT NOT NULL,
   quantity INT NOT NULL,
@@ -162,7 +162,7 @@ CREATE TABLE order_item (
 
 -- Pagamento (1:1 com pedido)
 CREATE TABLE payment (
-  id BIGINT AUTO_INCREMENT PRIMARY KEY,
+  id INT AUTO_INCREMENT PRIMARY KEY,
   order_id BIGINT UNIQUE NOT NULL,
   method VARCHAR(64) NOT NULL,
   amount DECIMAL(12,2) NOT NULL,
@@ -173,7 +173,7 @@ CREATE TABLE payment (
 
 -- Remessa (Shipment)
 CREATE TABLE shipment (
-  id BIGINT AUTO_INCREMENT PRIMARY KEY,
+  id INT AUTO_INCREMENT PRIMARY KEY,
   order_id BIGINT NOT NULL,
   carrier VARCHAR(128) NOT NULL,
   tracking_code VARCHAR(64) NOT NULL,
